front-end/requirements.txt
streamlit==1.53.0
pandas==2.2.3
numpy==1.26.4
plotly==5.19.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-dotenv==1.2.1
wordcloud==1.9.3
matplotlib==3.8.4
st-clickable-images
